mmc: ftsdc010_mci: Support DTS of ftsdc010 driver for generic dma
The ftsdc010 driver has been implemented for generic dma in Linux
kernel. And its compatible is andestech,atfsdc010g to distinguish
the legacy andestech,atfsdc010 which is not for generic dma.

Althought the ftsdc010_mci driver in U-Boot does not use dma, but
it still can work well with the mmc node for generic dma. So add
the compatible string to support it.

Signed-off-by: Rick Chen <rick@andestech.com>

Upstream-Status: Pending

Signed-off-by: Yu Chien Peter Lin <peterlin@andestech.com>